Stream Parser: Add basic support for BucketIndex messages
This doesn't bring the messages into the core of the app but it does add
support to the parser for recognizing the messages and partialy decoding
them. I haven't settled on the right internal representation for a
bucket index and I haven't decoded the index inside the response, but
this is just the first step.